--- frontend/components/receipts/error-boundary/index.tsx
@@ -1,1 +1,0 @@
-export * from './error-boundary'
