+++ frontend/components/receipts/receipt-error-boundary/index.tsx
@@ -1,0 +1,1 @@
+export * from './receipt-error-boundary'
