+++ frontend/components/receipts/receipt-error-boundary/receipt-error-boundary.tsx
